OMAP3+: voltage: expose transdone APIs
Transdone event in Voltage processor gives us fine grained status on
the current status of the voltage communication with the PMIC.
Unfortunately, IRQ generation by VP is based on the start of the
transmission to VC from VP, not at the end (or the completion of
the voltage setting). Hence any users of voltage layer who need
to know fine grained information such as confirmation if the voltage
is actually send to PMIC, needs to depend on this status.

Change-Id: I5fdef6ff9a6d4b90a5f04c3adfb4209c75b1a26e
Signed-off-by: Nishanth Menon <nm@ti.com>