OMAP4: PM: VC: support configuration of i2c clks
Patch "OMAP2+: voltage: split voltage controller (VC) code into dedicated layer"
splits out the hardcoded value in the code to vc's channel init.

This patch further isolates the configuration to remove out PMIC specific
configuration as high and low times are pmic specific.

Values are updated as well based on latest TI analysis done in android k35
kernel.

As discussed in http://marc.info/?l=linux-omap&m=130569587107997&w=2
we should, if possible isolate the changes away into board, PMIC and OMAP
specific parameters, to do this, we need a better understanding of the equations
involved, which we dont at this point of time and have to depend on emperical
values. This patch however is necessary instead of the hardcoded assumptions
in place.

Signed-off-by: Nishanth Menon <nm@ti.com>