TEMP: omap4: panda: Remove WLAN device registeration.
On Panda, we are observing that multiple suspend resume fails. On checking the
register dumps, this is narrowed down to UART2, which is stuck in transtion
and prevent idle. On Panda, WLAN is using UART2 interface. Removing the WLAN
device unblocks and allows CORE RET working multiple times. This is only ATM
and once issue is rootcaused with WLAN, this will be removed.

Signed-off-by: Govind Raj <govindraj.raja@ti.com>
Signed-off-by: Avinash.H.M <avinashhm@ti.com>