OMAP: PM: Rename omap-pm-noop.c to omap-pm-interface.c
Majority of the APIs available in omap-pm-noop.c and omap-pm.c are unused
and are deprecated. We retain the header compatibility with upstream, however,
the functionality in OMAP PM layer interfaces need to be retained. To prevent
code duplication and associated maintenance burden, we rename the current
omap-pm-noop as omap-pm-interface and add the deprecation and delta
functionality in subsequent patches

Acked-by: Nishanth Menon <nm@ti.com>
Acked-by: Ambresh K <ambresh@ti.com>,
Acked-by: Axel Haslam <axelhaslam@ti.com>
Signed-off-by: Girish S G <girishsg@ti.com>