OMAP: PM: merge omap-pm-interface and omap-pm files
Since we now have a lot of redundant code between omap-pm.c
and omap-pm-interface.c, we can conslidate the two and use
omap-pm-interface.c as the mainfile. omap-pm.c now contains
only the functionality needed when CONFIG_OMAP_PM=y.
A shared header is introduced to ensure that the function
prototypes remain consistent.

To maintain git history, we will rename omap-pm.c to omap-pm-helper.c
in a subsequent patch

[nm@ti.com: helped split the original patch]
Signed-off-by: Nishanth Menon <nm@ti.com>
Signed-off-by: Girish S G <girishsg@ti.com>