OMAP: PM: force omap_pm_get_device_context_loss_count
omap_pm_get_device_context_loss_count today tries to use the omap_device
based context loss logic to report back to the driver in-order for the
driver to optimally implement a save and restore. However,
as discussed in:
http://marc.info/?l=linux-omap&m=130814255217041&w=2
This is fraught with risk that the drivers at times may never get
reliable values. Meanwhile, while the upstream discusses this, we
choose the least risky path of reporting an incrementing context
loss forcing the drivers to maybe unoptimally keep restoring context.
This we ensure by preventing the rollover from ever reaching 0.

[nm@ti.com: helped split the original patch]
Signed-off-by: Nishanth Menon <nm@ti.com>
Signed-off-by: Girish S G <girishsg@ti.com>