tuna: sensors: Restore wakelock to proximity sensor
Restore the wakelock in a separate patch so it can be removed when
user land is repaired.

Arve Hjønnevåg: I looked some more at the user-space code, and
it is a lot of work to fix it. It puts the events back in a pipe
to deliver it to the apps, and since the user-space c wakelock
api does not support timeouts the workaround is easier to do in
the driver. If we need to wake up from this sensor, then lets
put the wakelock back in the driver as a separate patch
that can be reverted after the user-space code is fixed.

Change-Id: Ie427f00e594b3228d0b0b61481a011362e7cbcea
Signed-off-by: Marty Fouts <mfouts@sta.samsung.com>