HACK: OMAP4: hwmod: disable lcd1/lcd2/tv before reparenting dss clocks
HACK: There should be a way to do this without resorting to messing
with DISPC registers inside DSS reset func.

If we try to reparent the clocks while the LCD/TV is active and
in the middle of a DMA, we can get underflow/sync errors that
are sometimes unrecoverable. So, disable all the outputs first,
then reset DSS clock muxes to all point to PRCM clocks.

Based on a patch by Gilles-Arnaud Bleu-Laine <gilles@ti.com>

Change-Id: I89b39845d76bcb4c522a3b9fa7e95c7ee362c759
Signed-off-by: Dima Zavin <dima@android.com>