ARM: OMAP4: tuna: update power map for 4430 Tunas
The default TI configurations for 4430 uses OMAP4430+TWL6030.
on Tuna device, we need to have a configuration for OMAP4430+TPS+TWL
inline with what is present on 4460+TPS+TWL device.

Since pre-lunchboxes(4430) are no longer supported, we drop
the check for these and enable the map if 4430 device is detected.

Change-Id: Ic82ad411719c841af34b39ba8812d16c8a32e978
Signed-off-by: Nishanth Menon <nm@ti.com>
Signed-off-by: Vikram Pandita <vikram.pandita@ti.com>