OMAP4: PM: Disable SR in idle
SmartReflex values are not defined for retention voltage. Further, with
SmartReflex enabled, if CPU enters lower c-states, FSM will try to bump
the voltage to current OPP's voltage for which it has entered c-state;
hence SmartReflex needs to be disabled for MPU, CORE and IVA voltage
domains in idle path before enabling auto retention voltage achievement
on the device.

However, we could, in theory, hit a race condition between dvfs path
and idle path for a voltage domain, if detected, we reprogram our
next state to go to a power state which does not require SmartReflex
to be toggled around - basically give a priority for DVFS path on
conflict.

Acked-by: Todd Poynor <toddpoynor@google.com>
Acked-by: Santosh Shilimkar <santosh.shilimkar@ti.com>
[nm@ti.com: added dvfs check logic]
Signed-off-by: Nishanth Menon <nm@ti.com>
Signed-off-by: Ambresh K <ambresh@ti.com>
Signed-off-by: Rajendra Nayak <rnayak@ti.com>