OMAP3+: VC: calculate the setup times dynamically
To scale to the voltage during states such as retention, the PMIC requires some
voltage stabilization time. This depends on the worst case ramp voltage and the
sysclk. Support is added to dynamically calculate the ramp counts and update
the appropriate registers due to the impact of Errata Id: i623(Retention/Sleep
Voltage Transitions Ramp Time) on OMAP4430 silicon alone. This errata states
that the ramp time configuration in the PRM_VOLTSETUP_<VD name>_OFF and
PRM_VOLTSETUP_<VD name>_RET_SLEEP is 64, 256, 512, 2048 instead of the
expected 16, 64, 128, 512 sysclk cycles for the values 0, 1, 2 3 configured in
the register field. Since sysclk is a variant on OMAP4, dynamic computation
allows us to support multiple silicon variants runtime.

Acked-by: Todd Poynor <toddpoynor@google.com>
Signed-off-by: Ambresh K <ambresh@ti.com>
[Ported to 3.0 and cleanup]
Signed-off-by: Avinash.H.M. <avinashhm@ti.com>
Signed-off-by: Rajendra Nayak <rnayak@ti.com>
Signed-off-by: Nicole Chalhoub <n-chalhoub@ti.com>