ASoC: sdp4430: Configure modem's McBSP port for BT call
Configuration of McBSP port connected to modem requires a
special handling because it's not associated to any backend
but to the ABE modem frontend. Backends need to explicitly
configure modem's McBSP port, it was done for PDM backend
but was missing for McBSP backend which is used during BT
call.

Change-Id: Icb433a09ebe868e148933bbc5df7fd5ed0bda83a
Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>