OMAP3+: VC: add ability for auto transition
Auto transition is an ability on OMAP3+ devices with SmartReflex
capability to be able to automatically command a specific voltage
in reaction to a device/domain wide power state transition. In order
to support this variance in OMAP3 and OMAP4 devices, we provide a
new API to allow SoC specific OMAP PM code to program required
domain transition as they are detected.

Signed-off-by: Nishanth Menon <nm@ti.com>